requirements.txt
# ============================================================
# Incident Triage Copilot — Dependencies
# ============================================================
#
# WHY EACH PACKAGE EXISTS:
#
# mcp            — Official Model Context Protocol SDK.
#                  Provides FastMCP (server), ClientSession (client),
#                  and transport layers (stdio, SSE).
#
# pydantic       — Data validation & serialization.
#                  Powers our "incident contract" — strict types
#                  for inputs, outputs, evidence, mitigations.
#
# openai         — LLM client (works with OpenAI, Azure OpenAI,
#                  or any OpenAI-compatible endpoint).
#                  The agent's "brain" that reasons over tools.
#
# python-dotenv  — Loads .env files so we never hardcode secrets.
#
# rich           — Beautiful terminal output for the demo.
#                  Makes triage reports readable in the console.
#
# httpx          — Async HTTP client. Used by MCP transports
#                  and would be used for real API calls.
#
# uvicorn        — ASGI server, needed if we expose MCP servers
#                  over SSE (HTTP) instead of stdio.
# ============================================================

mcp>=1.2.0
pydantic>=2.0
openai>=1.12.0
python-dotenv>=1.0.0
rich>=13.0
httpx>=0.27.0
uvicorn>=0.29.0
